+++ src/modules/expense/components/ExpenseRoutingComponents.tsx
@@ -1,4 +1,4 @@
-import { ApprovalDetail } from '@expense/components/approval/detail/ApprovalDetail';
+import { ExpenseApprovalDetail } from '@expense/components/approval/detail/ExpenseApprovalDetail';
 import { ExpenseApprovalListView } from '@expense/components/approval/list/ExpenseApprovalListView';
 import { ExpenseRequestDetail } from '@expense/components/request/detail/ExpenseRequestDetail';
 import RequestEditor from '@expense/components/request/editor/RequestEditor';
@@ -7,21 +7,21 @@ import { Layout } from '@layout/components/base';
 import * as React from 'react';
 import { Route, RouteComponentProps, Switch } from 'react-router';
 
-const RequestList = () => <ExpenseRequestListView />;
-const ApprovalList = () => <ExpenseApprovalListView />;
+const ExpenseRequestList = () => <ExpenseRequestListView />;
+const ExpenseApprovalList = () => <ExpenseApprovalListView />;
 
 const request = (props: RouteComponentProps) => (
   <Switch>
     <Route path={`${props.match.path}/form`} component={RequestEditor} />
     <Route path={`${props.match.path}/:expenseUid`} component={ExpenseRequestDetail} />
-    <Route path={`${props.match.path}`} component={RequestList} />
+    <Route path={`${props.match.path}`} component={ExpenseRequestList} />
   </Switch>
 );
 
 const approval = (props: RouteComponentProps) => (
   <Switch>
-    <Route path={`${props.match.path}/:expenseUid`} component={ApprovalDetail} />
-    <Route path={`${props.match.path}`} component={ApprovalList} />
+    <Route path={`${props.match.path}/:expenseUid`} component={ExpenseApprovalDetail} />
+    <Route path={`${props.match.path}`} component={ExpenseApprovalList} />
   </Switch>
 );
