--- src/modules/expense/components/approval/detail/ApprovalDetail.ts
@@ -1,269 +1,0 @@
-import AppMenu from '@constants/AppMenu';
-import { IExpenseApprovalPostPayload } from '@expense/classes/request/approval';
-import { WithExpenseApproval, withExpenseApproval } from '@expense/hoc/withExpenseApproval';
-import { WithAppBar, withAppBar } from '@layout/hoc/withAppBar';
-import { WithLayout, withLayout } from '@layout/hoc/withLayout';
-import { WithUser, withUser } from '@layout/hoc/withUser';
-import { InjectedIntlProps, injectIntl } from 'react-intl';
-import { RouteComponentProps, withRouter } from 'react-router';
-import {
-  compose,
-  HandleCreators,
-  lifecycle,
-  mapper,
-  ReactLifeCycleFunctions,
-  StateHandler,
-  StateHandlerMap,
-  StateUpdaters,
-  withHandlers,
-  withStateHandlers,
-} from 'recompose';
-import { Dispatch } from 'redux';
-import { FormErrors } from 'redux-form';
-import { isNullOrUndefined, isObject } from 'util';
-
-import { WorkflowStatusType } from '@common/classes/types';
-import { ExpenseUserAction } from '@expense/classes/types';
-import { expenseMessages } from '@expense/locales/messages/expenseMessages';
-import { RadioGroupChoice } from '@layout/components/input/radioGroup';
-import { IAppBarMenu } from '@layout/interfaces';
-import { layoutMessage } from '@layout/locales/messages';
-import { WorkflowApprovalFormData } from '@organization/components/workflow/approval/WorkflowApprovalForm';
-import { organizationMessage } from '@organization/locales/messages/organizationMessage';
-import { ApprovalDetailView } from './ApprovalDetailView';
-
-interface OwnHandlers {
-  handleRefresh: () => void;
-  handleValidate: (payload: WorkflowApprovalFormData) => FormErrors;
-  handleSubmit: (payload: WorkflowApprovalFormData) => void;
-  handleSubmitSuccess: (result: any, dispatch: Dispatch<any>) => void;
-  handleSubmitFail: (errors: FormErrors | undefined, dispatch: Dispatch<any>, submitError: any) => void;
-}
-
-interface OwnRouteParams {
-  expenseUid: string;
-}
-
-interface OwnState {
-  isApprove?: boolean | undefined;
-  approvalTitle: string;
-  approvalSubHeader: string;
-  approvalChoices: RadioGroupChoice[];
-  approvalTrueValue: string;
-  approvalDialogTitle: string;
-  approvalDialogContentText: string;
-  approvalDialogCancelText: string;
-  approvalDialogConfirmedText: string;
-}
-
-interface OwnStateUpdaters extends StateHandlerMap<OwnState> {
-  stateUpdate: StateHandler<OwnState>;
-}
-
-export type ExpenseApprovalDetailProps
-  = WithExpenseApproval
-  & WithUser
-  & WithLayout
-  & WithAppBar
-  & RouteComponentProps<OwnRouteParams>
-  & InjectedIntlProps
-  & OwnHandlers
-  & OwnState
-  & OwnStateUpdaters;
-
-const handlerCreators: HandleCreators<ExpenseApprovalDetailProps, OwnHandlers> = {
-  handleValidate: (props: ExpenseApprovalDetailProps) => (formData: WorkflowApprovalFormData) => { 
-    const errors = {
-      information: {}
-    };
-  
-    const requiredFields = ['isApproved', 'remark'];    
-  
-    requiredFields.forEach(field => {
-      if (!formData[field] || isNullOrUndefined(formData[field])) {
-        errors.information[field] = props.intl.formatMessage(organizationMessage.workflow.fieldFor(field, 'fieldRequired'));
-      }
-    });
-    
-    return errors;
-  },
-  handleSubmit: (props: ExpenseApprovalDetailProps) => (formData: WorkflowApprovalFormData) => { 
-    const { match, intl, stateUpdate } = props;
-    const { user } = props.userState;
-    const { createRequest } = props.expenseApprovalDispatch;
-
-    // user checking
-    if (!user) {
-      return Promise.reject('user was not found');
-    }
-
-    // props checking
-    if (!match.params.expenseUid) {
-      const message = intl.formatMessage(expenseMessages.request.message.emptyProps);
-
-      return Promise.reject(message);
-    }
-
-    // compare approval status string
-    const isApproved = formData.isApproved === WorkflowStatusType.Approved;
-
-    stateUpdate({
-      isApprove: isApproved,
-    });
-
-    // generate payload
-    const payload: IExpenseApprovalPostPayload = {
-      isApproved,
-      remark: !isApproved ? formData.remark : null
-    };
-
-    // dispatch create request
-    return new Promise((resolve, reject) => {
-      createRequest({
-        resolve, 
-        reject,
-        expenseUid: match.params.expenseUid, 
-        companyUid: user.company.uid,
-        positionUid: user.position.uid,
-        data: payload, 
-      });
-    });
-  },
-  handleSubmitSuccess: (props: ExpenseApprovalDetailProps) => (response: boolean) => {
-    const { intl, history, match, isApprove } = props;
-    const { alertAdd } = props.layoutDispatch;
-
-    let message: string = '';
-
-    if (isApprove) {
-      message = intl.formatMessage(expenseMessages.approval.message.approveSuccess, {uid: match.params.expenseUid});
-    } else {
-      message = intl.formatMessage(expenseMessages.approval.message.rejectSuccess, {uid: match.params.expenseUid});
-    }
-
-    alertAdd({
-      message,
-      time: new Date()
-    });
-
-    if (match.params.expenseUid) {
-      history.push(`/expense/approvals`);
-    }
-  },
-  handleSubmitFail: (props: ExpenseApprovalDetailProps) => (errors: FormErrors | undefined, dispatch: Dispatch<any>, submitError: any) => {
-    const { intl } = props;
-    const { alertAdd } = props.layoutDispatch;
-    
-    if (errors) {
-      // validation errors from server (400: Bad Request)
-      alertAdd({
-        time: new Date(),
-        message: isObject(submitError) ? submitError.message : submitError
-      });
-    } else {
-      alertAdd({
-        time: new Date(),
-        message: intl.formatMessage(expenseMessages.approval.message.createFailure),
-        details: isObject(submitError) ? submitError.message : submitError
-      });
-    }
-  },
-  handleRefresh: (props: ExpenseApprovalDetailProps) => () => { 
-    const { match } = props;
-    const { user } = props.userState;
-    const { loadDetailRequest } = props.expenseApprovalDispatch;
-
-    if (user) {
-      loadDetailRequest({
-        companyUid: user.company.uid,
-        positionUid: user.position.uid,
-        expenseUid: match.params.expenseUid
-      });
-    }
-  }
-};
-
-const stateUpdaters: StateUpdaters<{}, OwnState, OwnStateUpdaters> = {
-  stateUpdate: (prevState: OwnState) => (newState: any) => ({
-    ...prevState,
-    ...newState
-  })
-};
-
-const lifecycles: ReactLifeCycleFunctions<ExpenseApprovalDetailProps, {}> = {
-  componentDidMount() {
-    const { layoutDispatch, intl, appBarDispatch, handleRefresh, match } = this.props;
-    const { user } = this.props.userState;
-    const { loadDetailRequest } = this.props.expenseApprovalDispatch;
-
-    layoutDispatch.changeView({
-      uid: AppMenu.ExpenseApproval,
-      parentUid: AppMenu.Expense,
-      title: intl.formatMessage(expenseMessages.approval.page.title),
-      subTitle : intl.formatMessage(expenseMessages.approval.page.subTitle)
-    });
-    
-    layoutDispatch.navBackShow();
-    
-    const handleMenuClick = (menu: IAppBarMenu): void => {
-      switch (menu.id) {
-        case ExpenseUserAction.Refresh:
-          handleRefresh();
-          break;
-
-        default:
-          break;
-      }
-    };
-
-    appBarDispatch.assignCallback(handleMenuClick);
-
-    if (user) {
-      loadDetailRequest({
-        companyUid: user.company.uid,
-        positionUid: user.position.uid,
-        expenseUid: match.params.expenseUid
-      });
-    }
-  },
-  componentWillUnmount() {
-    const { layoutDispatch, appBarDispatch } = this.props;
-
-    layoutDispatch.changeView(null);
-    layoutDispatch.navBackHide();
-    layoutDispatch.moreHide();
-    layoutDispatch.actionCentreHide();
-
-    appBarDispatch.dispose();
-  }
-};
-
-const createProps: mapper<ExpenseApprovalDetailProps, OwnState> = (props: ExpenseApprovalDetailProps): OwnState => {
-  const { intl } = props;
-
-  return {
-    approvalTitle: intl.formatMessage(expenseMessages.approval.section.title),
-    approvalSubHeader: intl.formatMessage(expenseMessages.approval.section.subTitle),
-    approvalChoices: [
-      { value: WorkflowStatusType.Approved, label: intl.formatMessage(organizationMessage.workflow.option.approve) },
-      { value: WorkflowStatusType.Rejected, label: intl.formatMessage(organizationMessage.workflow.option.reject) }
-    ],
-    approvalTrueValue: WorkflowStatusType.Approved,
-    approvalDialogTitle: intl.formatMessage(expenseMessages.approval.dialog.title),
-    approvalDialogContentText: intl.formatMessage(expenseMessages.approval.dialog.content),
-    approvalDialogCancelText: intl.formatMessage(layoutMessage.action.cancel),
-    approvalDialogConfirmedText: intl.formatMessage(layoutMessage.action.continue),
-  };
-};
-
-export const ApprovalDetail = compose<ExpenseApprovalDetailProps, {}>(
-  withUser,
-  withLayout,
-  withAppBar,
-  withRouter,
-  withExpenseApproval,
-  injectIntl,
-  withStateHandlers<OwnState, OwnStateUpdaters, {}>(createProps, stateUpdaters),
-  withHandlers<ExpenseApprovalDetailProps, OwnHandlers>(handlerCreators),
-  lifecycle<ExpenseApprovalDetailProps, {}>(lifecycles),
-)(ApprovalDetailView);
